Update projekt_1.py
did rest of the countings. Maybe I will use sum with for instead of lenght of list - seems more logical. I realized when doing numerics count. With list there I would have to use most probably other for to count the stuff in the list. So outcome from this is that I should use sum everywhere to save time